+++ src/scdcb_metrics/_settings.py
@@ -5,7 +5,7 @@ from typing import Literal
 from rich.console import Console
 from rich.logging import RichHandler
 
-scib_logger = logging.getLogger("scib_metrics")
+scib_logger = logging.getLogger("scdcb_metrics")
 
 
 class ScibConfig:
@@ -15,12 +15,12 @@ class ScibConfig:
     --------
     To set the progress bar style, choose one of "rich", "tqdm"
 
-    >>> scib_metrics.settings.progress_bar_style = "rich"
+    >>> scdcb_metrics.settings.progress_bar_style = "rich"
 
     To set the verbosity
 
     >>> import logging
-    >>> scib_metrics.settings.verbosity = logging.INFO
+    >>> scdcb_metrics.settings.verbosity = logging.INFO
     """
 
     def __init__(
@@ -59,13 +59,13 @@ class ScibConfig:
     def verbosity(self, level: str | int):
         """Set verbosity level.
 
-        If "scib_metrics" logger has no StreamHandler, add one.
+        If "scdcb_metrics" logger has no StreamHandler, add one.
         Else, set its level to `level`.
 
         Parameters
         ----------
         level
-            Sets "scib_metrics" logging level to `level`
+            Sets "scdcb_metrics" logging level to `level`
         force_terminal
             Rich logging option, set to False if piping to file output.
         """
@@ -83,7 +83,7 @@ class ScibConfig:
             scib_logger.setLevel(level)
 
     def reset_logging_handler(self) -> None:
-        """Reset "scib_metrics" log handler to a basic RichHandler().
+        """Reset "scdcb_metrics" log handler to a basic RichHandler().
 
         This is useful if piping outputs to a file.