+++ src/scdcb_metrics/utils/_utils.py
@@ -9,7 +9,7 @@ from scipy.sparse import csr_matrix
 from sklearn.neighbors import NearestNeighbors
 from sklearn.utils import check_array
 
-from scib_metrics._types import ArrayLike, IntOrKey, NdArray
+from scdcb_metrics._types import ArrayLike, IntOrKey, NdArray
 
 
 def get_ndarray(x: ArrayDevice) -> np.ndarray: